+++ CDrawGraphReserve.h
@@ -2,7 +2,7 @@
 #include "IRenderReserve.h"
 class CDrawGraphReserve : public IRenderReserve {
  public:
-  CDrawGraphReserve(int GHandle, int x, int y, double Priority, int width,
+  CDrawGraphReserve(int GHandle, int x, int y, int Priority, int width,
                     int height);
   void Render() const;
 };
