+++ CDrawRotaGraphFwithBlendReserve.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 
 CDrawRotaGraphFwithBlendReserve::CDrawRotaGraphFwithBlendReserve(
-    int GHandle, double Priority, float x, float y, float angle, float extend,
+    int GHandle, int Priority, double x, double y, float angle, float extend,
     int color, int Blendmode, int value, int width, int height)
     : IRenderReserve(
           GHandle, Priority, x, y,
@@ -20,7 +20,7 @@ CDrawRotaGraphFwithBlendReserve::CDrawRotaGraphFwithBlendReserve(
 void CDrawRotaGraphFwithBlendReserve::Render() const {
   SetDrawBright((Color >> 16) & 0xFF, (Color >> 8) & 0xFF, (Color >> 0) & 0xFF);
   SetDrawBlendMode(Blendmode, value);
-  DrawRotaGraphF(x, y, extend, angle, GHandle, TRUE);
+  DrawRotaGraphF(float(x), float(y), extend, angle, GHandle, TRUE);
   SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
   SetDrawBright(0xFF, 0xFF, 0xFF);
 }
