+++ CDrawRotaGraphFwithBlendReserve.h
@@ -6,8 +6,8 @@ class CDrawRotaGraphFwithBlendReserve : public IRenderReserve {
   int Color, Blendmode, value;
 
  public:
-  CDrawRotaGraphFwithBlendReserve(int GHandle, double Priority, float x,
-                                  float y, float angle, float extend, int color,
+  CDrawRotaGraphFwithBlendReserve(int GHandle, int Priority, double x,
+                                  double y, float angle, float extend, int color,
                                   int Blendmode, int value, int width,
                                   int height);
   void Render() const;
