+++ CEffect_Bat_Swing.h
@@ -1,12 +1,12 @@
 #pragma once
 #include "CEffect.h"
 class CEffect_Bat_Swing : public CEffect {
-  double angle;
+  float angle;
   int cnt = 0;
   int Color;
 
  public:
-  CEffect_Bat_Swing(CVector position, double angle);
+  CEffect_Bat_Swing(CVector position, float angle);
 
   void Update();
   void Render() const;
