+++ CNumberDrawer.h
@@ -5,6 +5,6 @@ class CNumberDrawer {
  public:
   CNumberDrawer();
   void Draw(int x, int y, double num, int type, int style,
-            double priority) const;
-  void Draw(int x, int y, int num, int type, int style, double priority) const;
+            int priority) const;
+  void Draw(int x, int y, int num, int type, int style, int priority) const;
 };
